View & Orientation Properties Fix
Didn't actually demo the changes from the previous few commits. I derped and realized I didn't properly hook up the new signals with the old assignment to data attributes. I realize now that the data attributes themselves were what I was using to store the state, not simply reflect it to the DOM itself. It *was* the state, that's the thing. I forgot this part of the equation.