App Constant Split Up
Moving the 'app constant' namespace module things into their own files. There is so much going on in this project it is almost hard to not find anything to work on. Maybe that's why it's so complex now. Once things are organized it might make more sense to start adding features.

I may just remove the certain features around editor tabs for the final migration steps, because they are holding up just rewriting it all together.