Reactive Renderer Editors
Okay, worked out some things with the SolidJS Playground. I think this is how I am going to implement the editor setup that I am looking to replace. It is fully state-driven, and it doesn't use any classes or internal signals. I have to still wrap my mind around reactive contexts for stores though, they are a bit more finneckey to me than how signals are. It's a bit more confusing what places are able to be reactive for properly access and or object referencing. Eg, you can't just log out the state object, you have to modify it in some way, like JSON.stringify() it, or map it into a new shape. Then Solid will know that you need to access it reactively.