Rename Editor Reimpl
Devin Townsend Project - Deconstruction.

I found this album at the perfect time. It's astounding.

I had already been slowly getting into it for about a year or so, I just didn't understand the premise of it, and I think the understanding I have started to understand in myself, that's what made things click for it.