Nullable Editor Close
And a renamed function in general. Realized I named it differently on accident.

People I know have for the first time every given me feedback that they like the music I released (Recreational Outrage). This is new territory now, I now see that the next big step for me is to start sharing my work, and to be proud of what I do. It's not necessarily that "it's work that needs showing", that's something else. It's more that, I need to not worry about whether to show it or not. I want to be able to make what I made, and work hard to make sure people can see it. That's something I haven't explored yet. I haven't had an audience before, nor people interested in what I make and what I will make into the future. I can embrace that. I have the tools available to me to make that happen. It's just me being able to use them that I have to embrace. That comes from inside.