Liquid Glass Icon
Okay, now there is a Liquid Glass icon for the app! Currently it is just a custom one that you can add to the app after you've already installed it. This folder/file/project here is for Icon Composer, Apple's new app for creating Liquid Glass icons. I very much appreciate that it is in JSON. Super awesome.

To use it you open this in that app, then create a legacy macOS pre-Tahoe icon image, then create an `.icns` file from that. Then you drag it onto the icon of STE when opened with the Info sheet (Cmd+I). It won't change colors with the themes sadly. I will look into getting proper maskable support for that kind of thing, as I think it looks like some other maskable PWA icons seem to work properly with the auto Liquid Glass composition thing. Like it isn't in a custom format like this, it is just the PWA maskable icon, and Liquid Glass seems to somehow work properly on it.