Liquid Glass ICNS Target
Okay, ended up adding the actual file to the repo, for ease of access and for you to use on your own if you would like to as well. One less step to need to do all of that work yourself. Might as well just make it easier, as it just "makes sense" lol. The file is kind of big though, agh. That's alright though, reproducibility is key. Like how they say that code easier to understand is better than super small code. I very much like that take as well.