Back to the Frayed Experiments of the Olden Days
Realizations while revisiting some of my old demos from early 2020, and commentary from those:

Pro tip, probably quickest to copy the code right into stedit.app to view these random snippets 😋

Come to think of it, that was kind of the purpose of STE back then, yet I didn't even really realize it. I wanted something like Codepen, but for local saving. Now I think it would be cool to have some sort of iteration too, maybe not full-blown Git, but something like that
Might look into that, could be cool to make a tool to migrate something like that to Git too, later down the line
So it could teach you the concepts of version control and web development, then you could move to the "big tools" once you felt up to it
Heck yeah!!!!
Okay, going to do that now I think
STE is going back to it's roots, from being a general-purpose text editor, into being a learning/teaching tool

Okay, I very much am needing to bring back the DIY spirit of coding, I think that's what made it fun back then
This project is stupid funny, I love it
You know when something is funny and you can't stop laughing? Man, it seems like that's hard to come by now being a "grown up" the last few years haha

Okay, back to "commit message me", hi 👋

Rediscovering what this playground STE used to foster in relation to my experiments and creativity, it makes sense why I made it the way I did back then, and I think it clears up the idea of where it needs to go into the future. I started to see it as being more of a browser-first multi-purpose text editor, like a mild VSCode competitor, rather than something that's meant for quick experiments, which was exactly what I used it for back then.

So I think that is the new goal with this. This is a new step back to that vision. It may be a bit late to the party, but it is clearer than ever! I might be in another two week phase of STE, but that's okay. At least it's a new outlook I haven't tried before!

Again, this makes a lot of sense. Because I was used to doing things on my own, I started to learn how to build the tools I would need to be able to do what I wanted to do. Hence, from the ground up. When I needed more from STE, I would add more features. When something wasn't working, I would debug it. If I needed a package, I would learn how to import it. Did I want code highlighting? Okay, let's add that. I learned so much from doing things like that. I used that same concepts to learn how to make music as well, and seemingly with the facets of maintainability from what I learned when doing coding just a few months earlier (these were on a similar timeline and trajectory, just in different parallel paths. Music seemingly is taking much much longer to conceptualize. I think I am reaching a certain point to it now though, in now that I am moving on to the production side of things, away from the pure creativeness. I still have fun with the creative part of it, but that's not the explicit focus as much anymore. I am looking to work on wrangling it back into the picture though, as it's necessary!)

Oh, and this file is the first of the new changes going forward with the "new old STE". It's not really related to this commit message specifically, other than it being related to the concept here at heart.