new editor Demo
Just for the concept of starting fresh, with familiar concepts and the original goal of the project, I made the remnants of the first version of STE, with Real-time HTML Editor.

https://htmledit.squarefree.com/

So, it once again now allows for itself to be coded by itself. I am debating whether maybe it may be a fun exercise to try doing this again in this manner, for the literal sense of not using all of the big tooling I do now. I very much learned a lot of my programming by being in the browser itself, because that's all I had available, back when I was on Chromebook.

So I kind of want to see what kinds of things I build now, with these more clean concepts in terms of organization and separation of concerns.

For complex projects, full developer tooling is indeed necessary. But I am indeed interested in going back to that original concept of just writing things small, as I did in the old days. The interesting part is that I wonder how much you can do, with using proper state management, and signals, but doing it by hand. I know I don't need to, but I want to keep clean code, AND have it all still be built using only the browser. Having it be open source of course is great too. I mean in that it's awesome to be able to easily explain how the site works, because of being able to edit the project with itself.

The bummer is that some things just aren't natural in this setup, hence why I moved over to a bundler, TypeScript, and a dev server.

Programming in the browser directly does feel more fresh than anything else can be though, and I am so glad I started with that in the beginning. Nothing beats it.

I do wish I could merely just have formatting, and syntax highlighting though. Having that without a bundler is a headache though. And not having autocomplete and type checking for APIs is less than ideal too. But for making concepts and things like that, plain web dev surely is goated, if I say so myself. I do agree that types do slow you down, albeit yes much safer and more secure code.

It is a trade of concerns for sure. I may want to try going back on this trade, just for a little while again. I want to rekindle that open trying of things again, and have a revised environment to reflect that. STE has been a bit straightforward and refined for a good while now, and I think it needs a change up. It's just that I feel like it lost it's roots. I want to bring them back. And I think it may have to come in the same fashion it was created originally. :)

But yeah, this file was all by hand, and by memory, to some extent. The only thing I googled was setSelectionRange, and I still got the syntax and parameters correct.

So I also do believe in knowing things outright. Web-stable APIs definitely are awesome to know by memory, I really did do hard work practicing that back then, I'm happy to have that as a foundation I can still use to this day.

Even having things like nesting nowadays really aids in the concept of making component-led CSS declarations too, which is great.

I'm curious what a "classic STE" with modern web features, and a domain would look like. Modern STE is already both of those, yet extremely and heavily fragmented because of the technical debt that occurred over time.

I'm tired of the old codebase. I successfully migrated it to SolidJS. I just don't like the codebase now, at all. Maybe because it's nothing like how it was constructed back early on, and it's just over-constructed for what it needs to be in the first place. It's a simple app. It doesn't need all of the tech of the world.