Opening with File System Access
Okay, using effects for file handles makes a lot more sense now. This is where state management was very tangled with the original project.

And if anything, maybe just learning how I want to structure things here, this can merely be brought to the existing codebase too. This is yet another rethinking of this original process, I think it is inevitable. Do I refactor? Do I rewrite? Do I replace? Do I keep it the same? Is it fine as-is? This metric is always changing. Either way, having a clean slate seems to help with seeing things from a new perspective. Requiring it to be the same as the original is only going to hinder things from being able to be implemented differently.